# contract/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Artwork": {
        "additionalProperties": false,
        "properties": {
          "accent": {
            "description": "Dominant vibrant colour as #rrggbb; - when the image yielded none.",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "height": {
            "description": "0 when not measured.",
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "poster",
              "backdrop",
              "thumb",
              "avatar",
              "banner"
            ],
            "type": "string"
          },
          "ownerId": {
            "description": "The owner's uuid, or the numeric user id for user artwork.",
            "type": "string"
          },
          "ownerKind": {
            "enum": [
              "title",
              "season",
              "episode",
              "user"
            ],
            "type": "string"
          },
          "path": {
            "description": "Location under the server's data directory.",
            "type": "string"
          },
          "source": {
            "enum": [
              "tmdb",
              "uploaded",
              "embedded"
            ],
            "type": "string"
          },
          "width": {
            "description": "0 when not measured.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "ownerKind",
          "ownerId",
          "kind",
          "path",
          "width",
          "height",
          "source",
          "createdAt"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerDevice": {
        "description": "Native clients: a device token from signInDevice, pollPairing or connectDevice.",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieSession": {
        "description": "Browsers.",
        "in": "cookie",
        "name": "couchverse_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Couchverse",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/artwork": {
      "post": {
        "operationId": "adminUploadArtwork",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "encoding": {
                "file": {
                  "contentType": "application/octet-stream"
                },
                "kind": {
                  "contentType": "text/plain"
                },
                "ownerId": {
                  "contentType": "text/plain"
                },
                "ownerKind": {
                  "contentType": "text/plain"
                }
              },
              "schema": {
                "properties": {
                  "file": {
                    "contentEncoding": "binary",
                    "contentMediaType": "application/octet-stream",
                    "description": "A .jpg, .jpeg, .png or .webp image (checked by extension).",
                    "format": "binary",
                    "type": "string"
                  },
                  "kind": {
                    "enum": [
                      "poster",
                      "backdrop",
                      "thumb"
                    ],
                    "type": "string"
                  },
                  "ownerId": {
                    "type": "string"
                  },
                  "ownerKind": {
                    "enum": [
                      "title",
                      "season",
                      "episode"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "ownerKind",
                  "ownerId",
                  "kind",
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artwork"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerDevice": []
          }
        ],
        "tags": [
          "artwork"
        ]
      }
    },
    "/admin/artwork/{id}": {
      "delete": {
        "operationId": "adminDeleteArtwork",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerDevice": []
          }
        ],
        "tags": [
          "artwork"
        ]
      }
    },
...
    "/artwork/{id}": {
      "get": {
        "description": "The stored original (JPEG, PNG or WebP), or with size a cached JPEG resize.",
        "operationId": "getArtwork",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Resize to this width; the original when omitted.",
            "in": "query",
            "name": "size",
            "schema": {
              "enum": [
                "w342",
                "w780"
              ],
              "type": "string"
            }
          },
          {
            "description": "Version token (the artwork's createdAt); a versioned URL is cached as immutable.",
            "in": "query",
            "name": "v",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "An artwork grant, for a request without the session (a system image fetch, an anonymous couch guest).",
            "in": "query",
            "name": "g",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "contentMediaType": "application/octet-stream",
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "contentMediaType": "application/octet-stream",
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "contentMediaType": "application/octet-stream",
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The image"
          },
          "304": {
            "description": "Not modified since If-Modified-Since"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get an artwork image",
        "tags": [
          "artwork"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}